requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
django-filter==23.5
gunicorn==20.1.0
python-dotenv==1.0.0
django-filter==23.5
setuptools==68.2.2
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==20.1.0
drf-spectacular==0.24.1
